Move constants out of config and consolidate config loading (#40)
I noticed that many files were opening the `config.yml` file and reading
the same values from there. I've moved this into `config.py` so that we
only read the config once.

I also saw that the `config.yml` file was being read from the current
working directory, but this will only work if you are in the
`rapids-cli` project directory. I've updated this to read the config
from a relative position so it will work from anywhere.

I also spotted that the emoji constants being used were stored in the
config. These should probably be stored in code so I've moved those to
`constants.py`.