Register checks via entrypoints (#41)
Registers all checks via entrypoints in `pyproject.toml`. Then replaces
the `default_checks()` function with a walk over the entrypoints.

It looks like the modern way to discover and iterate over entrypoints is
to use `importlib.metadata.entry_points` from the standard library.
However for Python 3.11 and below you need to use
`importlib_metadata.entry_points` which is a separate package. So I've
added that dependency for older Python versions and introduced a
`_compatibility.py` file with a small shim so that we are using the
right version for our Python version.

I then went through the `default_checks()` functions and made sure that
each one was self contained. Some of them took config values as
arguments. Instead I've made it so that each check looks up its own
config values at the time that it needs them.

Also some of the checks are skipped if other checks fail. With
entrypoints we wont be able to do this.

**Note: Entrypoints are only registered when you install a package, so
you will need to run `pip install -e .` whenever the
`[project.entry-points.rapids_doctor_check]` section of `pyproject.toml`
is modified.**