build and test wheels in CI (#95)
Fixes #90 

Replaces #93 *(was automatically closed by GitHub when we flipped the
repo from private to public)*

* adds wheels CI (for PRs and branch builds)
* changes the distribution name from `rapids_cli` (underscore) to
`rapids-cli` (hyphen), so the experience will be e.g. `pip install
rapids-cli`
* puts a floor on `setuptools`, to ensure builds always get a version
that supports `pyproject.toml`
* sets up tests to run on CI runners with GPUs
  - *I expect we'll want this as the test suite here is expanded*

## Notes for Reviewers

### Why use the RAPIDS calendar-versioned workflows?

I think it's useful here to reuse all the RAPIDS machinery, especially
artifact-handling and running tests on runners with GPUs.

This project differs in this important way from other pure-Python,
`main`-branch projects in RAPIDS like
https://github.com/rapidsai/rapids-build-backend or
https://github.com/rapidsai/rapids-metadata ... something like
https://github.com/rapidsai/dask-cuda is probably a better reference for
the testing needs for this project.

Maybe we'll want to revisit this in the future as we see how often this
project gets released relative to RAPIDS, but I think this is a good
starting point for now.